process: alomst done with the site
